+++ src/main/kotlin/hackathon/sprinter/mockpost/dto/MockPostDao.kt
@@ -2,7 +2,7 @@ package hackathon.sprinter.mockpost.dto
 
 import hackathon.sprinter.util.MockBaseEntity
 
-data class MockPostDAO(
+data class MockPostDao(
     val email: String,
     val projectName: String,
     val link: String,
