+++ src/main/kotlin/hackathon/sprinter/mockpost/repository/MockPostRepositoryV2.kt
@@ -2,7 +2,7 @@ package hackathon.sprinter.mockpost.repository
 
 import com.google.gson.Gson
 import com.google.gson.JsonObject
-import hackathon.sprinter.mockpost.dto.MockPostDAO
+import hackathon.sprinter.mockpost.dto.MockPostDao
 import hackathon.sprinter.mockpost.dto.MockPostInput
 import hackathon.sprinter.util.RestTemplateService
 import org.springframework.http.HttpMethod
@@ -13,7 +13,7 @@ class MockPostRepositoryV2(
     private val restTemplateService: RestTemplateService,
     private val gson: Gson,
 ) {
-    fun findMockPostList(page: Int, size: Int): List<MockPostDAO> {
+    fun findMockPostList(page: Int, size: Int): List<MockPostDao> {
         return restTemplateService.fetch(
             HttpMethod.GET,
             "/api/collections/mockposts/records?page=$page&perPage=$size",
@@ -21,10 +21,10 @@ class MockPostRepositoryV2(
             null,
             "",
         ).get("items").asJsonArray
-            .let { it.map { gson.fromJson(it, MockPostDAO::class.java) } }
+            .let { it.map { gson.fromJson(it, MockPostDao::class.java) } }
     }
 
-    fun findMockPostById(id: String): MockPostDAO {
+    fun findMockPostById(id: String): MockPostDao {
         return restTemplateService.fetch(
             HttpMethod.GET,
             "/api/collections/mockposts/records?id=$id",
@@ -33,7 +33,7 @@ class MockPostRepositoryV2(
             "",
         ).get("items").asJsonArray.firstOrNull()
             .let { it ?: throw NoSuchElementException("No such element with id: $id") }
-            .let { gson.fromJson(it, MockPostDAO::class.java) }
+            .let { gson.fromJson(it, MockPostDao::class.java) }
     }
 
     fun createMockPost(mockPostInput: MockPostInput): String {
