+++ src/main/kotlin/hackathon/sprinter/mockpost/service/MockPostQueryService.kt
@@ -1,6 +1,6 @@
 package hackathon.sprinter.mockpost.service
 
-import hackathon.sprinter.mockpost.model.MockPostModel
+import hackathon.sprinter.mockpost.model.MockPostDto
 import hackathon.sprinter.mockpost.repository.MockPostRepositoryV2
 import org.springframework.stereotype.Service
 
@@ -8,13 +8,13 @@ import org.springframework.stereotype.Service
 class MockPostQueryService(
     private val mockPostRepositoryV2: MockPostRepositoryV2,
 ) {
-    fun findMockPostListByPaging(page: Int, size: Int): List<MockPostModel> {
+    fun findMockPostListByPaging(page: Int, size: Int): List<MockPostDto> {
         return mockPostRepositoryV2.findMockPostList(page, size)
-            .map { dao -> MockPostModel.from(dao) }
+            .map { dao -> MockPostDto.from(dao) }
     }
 
-    fun findMockPostById(id: String): MockPostModel {
+    fun findMockPostById(id: String): MockPostDto {
         return mockPostRepositoryV2.findMockPostById(id)
-            .let { dao -> MockPostModel.from(dao)}
+            .let { dao -> MockPostDto.from(dao)}
     }
 }
