+++ src/main/resources/application.yml
@@ -1,5 +1,5 @@
 server:
-  port: 9090
+  port: 8080
 
 spring:
   graphql:
